Move filters into useSearchType and use for type state
